fix: don't convert on basic types

+++ file/file_test.go
@@ -228,7 +228,7 @@ testData:
 		TestData: map[string]interface{}{
 			"Headers": map[string]interface{}{"Foo": "Bar"},
 			"trustIP": []interface{}{"10.0.0.0/8", "172.0.0.0/8", "192.0.0.0/8"},
-			"koo":     []interface{}{1, 2, 3},
+			"koo":     []interface{}{int64(1), int64(2), int64(3)},
 			"soo":     []interface{}{"1", "a", "3"},
 			"boo":     []interface{}{float64(1), 2.6, float64(3)},
 			"hoo":     []interface{}{true, false, false, true},
